CHANGES: - Send alert messages to audit channel, if no admin channel is available
BUGFIX:
- Bad missionLists could prevent a server start
- LogAnalyser: potential FileNotFound error solved
- Music: /music play <song> autocomplete broken

+++ core/data/impl/serverimpl.py
@@ -432,7 +432,7 @@ class ServerImpl(Server):
                 _mission = mission
             # check if the orig file has been updated
             orig = _mission + '.orig'
-            if os.path.exists(orig) and os.path.getmtime(orig) > os.path.getmtime(mission):
+            if os.path.exists(orig) and os.path.exists(mission) and os.path.getmtime(orig) > os.path.getmtime(mission):
                 shutil.copy2(orig, _mission)
                 missions.append(_mission)
             elif os.path.exists(mission):
@@ -681,37 +681,39 @@ class ServerImpl(Server):
                 if not filename:
                     self.log.warning("No mission found. Is your mission list empty?")
                     return filename
+
             new_filename = utils.get_orig_file(filename)
-            # process all mission modifications
-            dirty = False
-            for ext in self.extensions.values():
-                if type(ext).beforeMissionLoad != Extension.beforeMissionLoad:
-                    new_filename, _dirty = await ext.beforeMissionLoad(new_filename)
-                    if _dirty:
-                        self.log.info(f'  => {ext.name} applied on {new_filename}.')
-                    dirty |= _dirty
-            # we did not change anything in the mission
-            if not dirty:
+            try:
+                # process all mission modifications
+                dirty = False
+                for ext in self.extensions.values():
+                    if type(ext).beforeMissionLoad != Extension.beforeMissionLoad:
+                        new_filename, _dirty = await ext.beforeMissionLoad(new_filename)
+                        if _dirty:
+                            self.log.info(f'  => {ext.name} applied on {new_filename}.')
+                        dirty |= _dirty
+                # we did not change anything in the mission
+                if not dirty:
+                    return filename
+                # check if the original mission can be written
+                if filename != new_filename:
+                    missions: list[str] = self.settings['missionList']
+                    try:
+                        index = missions.index(filename) + 1
+                        await self.replaceMission(index, new_filename)
+                    except ValueError:
+                        # we should not be here, but just in case
+                        if new_filename not in missions:
+                            await self.addMission(new_filename)
+                return new_filename
+            except Exception as ex:
+                if isinstance(ex, UnsupportedMizFileException):
+                    self.log.error(ex)
+                else:
+                    self.log.exception(ex)
+                if filename != new_filename and os.path.exists(new_filename):
+                    os.remove(new_filename)
                 return filename
-            # check if the original mission can be written
-            if filename != new_filename:
-                missions: list[str] = self.settings['missionList']
-                try:
-                    index = missions.index(filename) + 1
-                    await self.replaceMission(index, new_filename)
-                except ValueError:
-                    # we should not be here, but just in case
-                    if new_filename not in missions:
-                        await self.addMission(new_filename)
-            return new_filename
-        except Exception as ex:
-            if isinstance(ex, UnsupportedMizFileException):
-                self.log.error(ex)
-            else:
-                self.log.exception(ex)
-            if filename != new_filename and os.path.exists(new_filename):
-                os.remove(new_filename)
-            return filename
         finally:
             # enable autoscan
             if self.locals.get('autoscan', False):
